Added a conditional before killing gconfd-2.  This is also going to be catalyst 1.1.5, so let's hope we don't find any more bugs, at least for this release.
git-svn-id: svn+ssh://svn.gentoo.org/var/svnroot/catalyst/trunk@542 d1e1f19c-881f-0410-ab34-b69fee027534

+++ targets/grp/grp.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 # Copyright 1999-2004 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo/src/catalyst/targets/grp/Attic/grp.sh,v 1.20 2005/01/29 14:54:31 wolf31o2 Exp $
+# $Header: /var/cvsroot/gentoo/src/catalyst/targets/grp/Attic/grp.sh,v 1.21 2005/01/30 00:48:03 wolf31o2 Exp $
 
 case $1 in
 	enter)
@@ -20,10 +20,13 @@ case $1 in
 	;;
 
 	preclean)
-		killall -9 gconfd-2
-		cp ${clst_sharedir}/targets/grp/grp-preclean-chroot.sh ${clst_chroot_path}/tmp
-		${clst_CHROOT} ${clst_chroot_path} /tmp/grp-preclean-chroot.sh || exit 1
-		rm -f ${clst_chroot_path}/tmp/grp-preclean-chroot.sh
+		if [ `pidof gconfd-2` ]
+		then
+			killall -9 gconfd-2
+		fi
+		#cp ${clst_sharedir}/targets/grp/grp-preclean-chroot.sh ${clst_chroot_path}/tmp
+		#${clst_CHROOT} ${clst_chroot_path} /tmp/grp-preclean-chroot.sh || exit 1
+		#rm -f ${clst_chroot_path}/tmp/grp-preclean-chroot.sh
 		exit 0
 	;;
