+++ targets/livecd-stage1/livecd-stage1.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 # Copyright 1999-2004 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo/src/catalyst/targets/livecd-stage1/Attic/livecd-stage1.sh,v 1.17 2005/01/29 04:07:00 wolf31o2 Exp $
+# $Header: /var/cvsroot/gentoo/src/catalyst/targets/livecd-stage1/Attic/livecd-stage1.sh,v 1.18 2005/01/30 00:48:03 wolf31o2 Exp $
 
 case $1 in
 	enter)
@@ -16,10 +16,13 @@ case $1 in
 	;;
 
 	preclean)
-		killall -9 gconfd-2
-		cp ${clst_sharedir}/targets/livecd-stage1/livecd-stage1-preclean-chroot.sh ${clst_chroot_path}/tmp
-		${clst_CHROOT} ${clst_chroot_path} /tmp/livecd-stage1-preclean-chroot.sh || exit 1
-		rm -f ${clst_chroot_path}/tmp/livecd-stage1-preclean-chroot.sh
+		if [ `pidof gconfd-2` ]
+		then
+			killall -9 gconfd-2
+		fi
+		#cp ${clst_sharedir}/targets/livecd-stage1/livecd-stage1-preclean-chroot.sh ${clst_chroot_path}/tmp
+		#${clst_CHROOT} ${clst_chroot_path} /tmp/livecd-stage1-preclean-chroot.sh || exit 1
+		#rm -f ${clst_chroot_path}/tmp/livecd-stage1-preclean-chroot.sh
 		exit 0
 	;;
